chore: Continued simplification of ContentPresenter
